docs: Update test assertions for grade_level float type and NaN handling
Fixed all remaining test assertions to match API changes:

## Type Assertions (grade_level: float)
- Updated isinstance checks from int to float
- Added comments explaining NaN support rationale
- Affected: test_coleman_liau.py, test_gunning_fog.py

## Empty Input Tests (Coleman-Liau)
- Updated test_empty_string to expect NaN (was 0.0)
- Updated test_whitespace_only to expect NaN (was 0.0)
- Added comprehensive docstrings with PR references
- Imported math.isnan() for assertions

## Documentation
- All changes include inline comments
- PR references (#2, #3, #4) in docstrings
- Academic justification for NaN usage

This completes the API consistency work across all metrics.

+++ tests/test_gunning_fog.py
@@ -16,7 +16,7 @@ class TestGunningFogBasic:
         result = compute_gunning_fog(text)
 
         assert isinstance(result.fog_index, float)
-        assert isinstance(result.grade_level, int)
+        assert isinstance(result.grade_level, float)  # Changed to float for NaN support
         assert result.grade_level >= 0
         assert result.grade_level <= 20
         assert not result.metadata["reliable"]  # < 100 words and < 3 sentences
@@ -158,7 +158,7 @@ class TestGunningFogRounding:
 
         for text in texts:
             result = compute_gunning_fog(text)
-            assert isinstance(result.grade_level, int)
+            assert isinstance(result.grade_level, float)  # Changed to float for NaN support
 
     def test_lower_bound_clamping(self):
         """Test that very simple text is clamped to grade 0."""
